refactor: splitted server tests into smaller files
