+++ test/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "max-len": "off"
+  }
+}
